CSCFAIRMETA-978: [FIX] Add download options to "Download All" button (#833)
