+++ etsin_finder/frontend/js/components/dataset/data/idaResourcesV2/index.jsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
+import styled from 'styled-components'
 import { observer } from 'mobx-react'
 import Translate from 'react-translate-component'
 import { faFile, faFolder } from '@fortawesome/free-solid-svg-icons'
@@ -13,6 +14,8 @@ import ErrorMessage from './errorMessage'
 import PackageModal from './packageModal'
 import ManualDownloadModal from './manualDownloadModal'
 import { Header, HeaderTitle, HeaderStats, HeaderButton } from '../common/dataHeader'
+import { SplitButtonContainer, MoreButton } from './splitButton'
+import FlaggedComponent from '../../../general/flaggedComponent'
 
 const downloadAll = identifier => {
   const handle = window.open(`/api/dl?cr_id=${identifier}`)
@@ -23,7 +26,7 @@ const downloadAll = identifier => {
 
 function IdaResources(props) {
   const { restrictions } = props.Stores.Access
-  let allowDownload =
+  const allowDownload =
     props.dataset.data_catalog.catalog_json.identifier !== 'urn:nbn:fi:att:data-catalog-pas' &&
     restrictions.allowDataIdaDownloadButton
 
@@ -62,7 +65,7 @@ function IdaResources(props) {
 
   let downloadFunc = () => downloadAll(props.dataset.identifier)
   const buttonProps = {}
-  let downloadAllText = 'dataset.dl.downloadAll'
+  const downloadAllText = 'dataset.dl.downloadAll'
 
   // Download full dataset package
   const action = getDownloadAction(props.dataset.identifier, null, Packages, Files)
@@ -73,14 +76,19 @@ function IdaResources(props) {
   buttonProps.attributes = {
     tooltip: action.tooltip,
   }
-  if (DatasetQuery.isDraft) {
-    allowDownload = false
-    downloadAllText = 'dataset.dl.downloadDisabledForDraft'
-  } else if (action.pending) {
-    downloadAllText = 'dataset.dl.packages.pending'
-  } else if (!action.available) {
-    downloadAllText = 'dataset.dl.packages.createForAll'
-  }
+  const { moreFunc, moreAriaLabel } = action
+
+  const downloadButton = (
+    <Translate
+      component={HeaderButton}
+      disabled={!allowDownload}
+      onClick={downloadFunc}
+      {...buttonProps}
+    >
+      <Translate content={downloadAllText} />
+      <Translate className="sr-only" content="dataset.dl.file_types.both" />
+    </Translate>
+  )
 
   return (
     <>
@@ -91,15 +99,20 @@ function IdaResources(props) {
           {totalSize ? ` (${sizeParse(totalSize)})` : null}
         </HeaderStats>
 
-        <Translate
-          component={HeaderButton}
-          disabled={!allowDownload}
-          onClick={downloadFunc}
-          {...buttonProps}
-        >
-          <Translate content={downloadAllText} />
-          <Translate className="sr-only" content="dataset.dl.file_types.both" />
-        </Translate>
+        <FlaggedComponent flag="DOWNLOAD_API_V2.OPTIONS" whenDisabled={downloadButton}>
+          <HeaderButtonSplit split={moreFunc}>
+            {downloadButton}
+            {moreFunc && (
+              <Translate
+                component={MoreButton}
+                color={buttonProps.color}
+                disabled={!allowDownload}
+                onClick={moreFunc}
+                attributes={{ 'aria-label': moreAriaLabel }}
+              />
+            )}
+          </HeaderButtonSplit>
+        </FlaggedComponent>
       </Header>
 
       <ErrorMessage error={Packages.error} clear={Packages.clearError} />
@@ -111,6 +124,13 @@ function IdaResources(props) {
     </>
   )
 }
+
+const HeaderButtonSplit = styled(SplitButtonContainer)`
+  width: 10.5em;
+  & > button:last-child {
+    margin: 0;
+  }
+`
 
 IdaResources.propTypes = {
   dataset: PropTypes.object.isRequired,
